Don't cast the load average to int32_t

+++ sen2agri-persistence/persistencemanager.cpp
@@ -778,11 +778,11 @@ void PersistenceManagerDBProvider::InsertNodeStatistics(const NodeStatistics &st
         query.bindValue(QStringLiteral(":swapTotalKb"), qlonglong{ statistics.swapTotalKb });
         query.bindValue(QStringLiteral(":swapUsedKb"), qlonglong{ statistics.swapUsedKb });
         query.bindValue(QStringLiteral(":loadAvg1"),
-                        static_cast<int16_t>(lrint(100.0 * statistics.loadAvg1)));
+                        static_cast<int32_t>(lrint(100.0 * statistics.loadAvg1)));
         query.bindValue(QStringLiteral(":loadAvg5"),
-                        static_cast<int16_t>(lrint(100.0 * statistics.loadAvg5)));
+                        static_cast<int32_t>(lrint(100.0 * statistics.loadAvg5)));
         query.bindValue(QStringLiteral(":loadAvg15"),
-                        static_cast<int16_t>(lrint(100.0 * statistics.loadAvg15)));
+                        static_cast<int32_t>(lrint(100.0 * statistics.loadAvg15)));
         query.bindValue(QStringLiteral(":diskTotalBytes"), qlonglong{ statistics.diskTotalBytes });
         query.bindValue(QStringLiteral(":diskUsedBytes"), qlonglong{ statistics.diskUsedBytes });
         query.bindValue(QStringLiteral(":timestamp"), QDateTime::currentDateTimeUtc());
